+++ src/main/kotlin/com/dsoftware/ghmanager/api/GithubApi.kt
@@ -64,7 +64,7 @@ object GithubApi : GithubApiRequests.Entity("/repos") {
     }
 
 
-    fun getWorkflowRunJobs(url: String) = get<WorkflowRunJobs>(
+    fun getJobsForWorkFlowRun(url: String): GithubApiRequest<WorkflowRunJobs> = get<WorkflowRunJobs>(
         url, "Get workflow-run jobs", pagination = GithubRequestPagination(1)
     )