+++ src/test/kotlin/com/dsoftware/ghmanager/GitHubActionsManagerBaseTest.kt
@@ -1,5 +1,6 @@
 package com.dsoftware.ghmanager
 
+import com.dsoftware.ghmanager.api.GhApiRequestExecutor
 import com.dsoftware.ghmanager.data.GhActionsService
 import com.dsoftware.ghmanager.ui.GhActionsMgrToolWindowContent
 import com.dsoftware.ghmanager.ui.settings.GhActionsSettingsService
@@ -18,6 +19,7 @@ import com.intellij.toolWindow.ToolWindowHeadlessManagerImpl
 import com.intellij.util.concurrency.annotations.RequiresEdt
 import com.intellij.util.ui.UIUtil
 import io.mockk.every
+import io.mockk.impl.annotations.MockK
 import io.mockk.mockk
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -62,8 +64,6 @@ abstract class GitHubActionsManagerBaseTest {
 
     @AfterEach
     open fun tearDown() {
-        val toolWindowManager = ToolWindowHeadlessManagerImpl(projectRule.project)
-        toolWindowManager.unregisterToolWindow("GitHub Actions")
         executeSomeCoroutineTasksAndDispatchAllInvocationEvents(projectRule.project)
         TestApplicationManager.tearDownProjectAndApp(projectRule.project)
     }