+++ CHANGELOG.md
@@ -4,6 +4,24 @@
 
 ## Unreleased
 
+## 1.21.1
+
+### 🐛 Bug Fixes
+
+- Storing action data in project settings instead of a separate file.
+
+## 1.21.0
+
+### 🚀 Features
+
+- Highlight outdated actions in workflow files. #122
+- Quick fix action to update to the latest major version of the action. #122
+- Error reporting directly to GitHub issues.
+
+### 🧰 Maintenance
+
+- Update bug-report template.
+
 ## 1.20.0
 
 ### 🚀 Features